% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTSS.R
\name{calculateTSS}
\alias{calculateTSS}
\title{A function to calculate TSS (total sum of squares)}
\usage{
calculateTSS(data, column)
}
\arguments{
\item{data}{a data frame}

\item{column}{name of the column to calculate TSS for}
}
\value{
an integer value of TSS
}
\description{
A function to calculate TSS (total sum of squares)
}
\examples{
calculateTSS(iris, "Sepal.Length")
}
