% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Confidence Interval Calculation with Histogram Plot}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Integer. Number of bootstrap iterations. Default is 10,000.}

\item{x}{Numeric vector. The sample data from which to bootstrap.}

\item{fun}{Character. The statistic to be calculated in each bootstrap sample. Default is `"mean"`.}

\item{alpha}{Numeric. Significance level for the confidence interval. Default is 0.05.}

\item{cx}{Numeric. The size of the text annotations in the plot. Default is 1.5.}

\item{...}{Additional graphical parameters to be passed to `hist()`.}
}
\value{
A list with the following elements:
\describe{
  \item{ci}{The confidence interval of the statistic based on the bootstrap resampling.}
  \item{fun}{The statistic function used (e.g., `"mean"`).}
  \item{x}{The original sample data.}
  \item{xstat}{A vector of bootstrap sample statistics.}
}
}
\description{
This function performs a bootstrap sampling procedure to estimate the confidence interval of a specified statistic (e.g., mean, median) from a given sample. It also generates a histogram of the bootstrap sample statistics.
}
\details{
The function performs `iter` bootstrap resamples of the input sample `x`, applying the function specified in `fun` to each sample. The output includes the confidence interval, the histogram of the bootstrap statistics, and the point estimate (calculated on the original sample).
}
\examples{
# Sample data
sample_data <- rnorm(100, mean = 5, sd = 2)
# Calculate bootstrap confidence interval for the mean
result <- myboot2(iter = 5000, x = sample_data, fun = "mean", alpha = 0.05)

}
