% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Compute a 95\% Confidence Interval for the Mean}
\usage{
myci(d)
}
\arguments{
\item{d}{Numeric vector of sample data.}
}
\value{
Numeric vector of length 2 with the lower and upper bounds of the confidence interval.
}
\description{
Calculates a 95\% confidence interval for the mean of a sample.
}
\examples{
sample_data <- c(5, 7, 8, 9, 10)
myci(sample_data)

}
