% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Central Limit Theorem Simulation}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{Integer. The number of observations per iteration (sample size).}

\item{iter}{Integer. The number of iterations (number of samples).}
}
\value{
A vector containing the means of the samples generated in each iteration.
It also generates a histogram of the sample means.
}
\description{
This function simulates the Central Limit Theorem (CLT) by generating random samples
from a uniform distribution between 0 and 5, and calculating the means of these samples
over multiple iterations.
}
\details{
The function generates \code{n} observations \code{iter} times from a uniform
distribution ranging from 0 to 5. For each sample, it computes the mean and then
plots a histogram of these means. This demonstrates the Central Limit Theorem (CLT),
as the distribution of sample means tends to be normal as the number of iterations
increases, regardless of the distribution of the original data.
}
\examples{
# Simulate CLT with sample size 10 and 10000 iterations
myclt(n = 10, iter = 10000)

}
