% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot a Normal Curve and Calculate the Area Under the Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the normal distribution.}

\item{sigma}{The standard deviation of the normal distribution.}

\item{a}{The value up to which the area under the curve is shaded and the probability is calculated.}
}
\value{
A list containing the mean (`mu`), standard deviation (`sigma`), and the
area (probability P(X ≤ a)) under the curve.
}
\description{
This function plots a normal distribution curve based on the specified mean (`mu`) and
standard deviation (`sigma`), shades the area under the curve from -∞ to `a`, and
calculates the probability P(X ≤ a) for the given normal distribution.
}
\examples{
# Plot the curve for a normal distribution with mean = 10, sigma = 5, and find the
# area (probability) up to x = 6
myncurve(mu = 10, sigma = 5, a = 6)

}
