% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysquare.R
\name{mysquare}
\alias{mysquare}
\title{Square Function}
\usage{
mysquare(x)
}
\arguments{
\item{x}{A numeric vector. Each element should be a quantitative value to be squared.}
}
\value{
A numeric vector with each element squared.
}
\description{
This function computes the square of each element in a numeric vector.
}
\examples{
# Square a vector of integers from 1 to 10
mysquare(x = 1:10)

# Square a vector of real numbers
mysquare(x = c(0.5, 1.5, 2.5))
}
